backend/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
pillow
pytesseract
python-multipart
google-generativeai
typing-extensions
dataclasses


